refactor: split browser proving and saved-ID storage into a follow-up PR
