+++ packages/zkpassport-sdk/src/popup/client.ts
@@ -21,8 +21,6 @@ export type PopupCallbacks = {
   onResult?: (result: Omit<PopupResult, "zkpassport" | "type">) => void
   onReject?: () => void
   onError?: (message: string) => void
-  onEnrollmentSaved?: () => void
-  onEnrollmentDeclined?: () => void
   // Fired when the user closes the popup before a result was produced
   onClose?: () => void
 }
@@ -124,12 +122,6 @@ export function openVerificationPopup(
         break
       case "error":
         callbacks.onError?.(data.message)
-        break
-      case "enrollment-saved":
-        callbacks.onEnrollmentSaved?.()
-        break
-      case "enrollment-declined":
-        callbacks.onEnrollmentDeclined?.()
         break
     }
   }
