+++ packages/zkpassport-sdk/src/popup/protocol.ts
@@ -23,7 +23,6 @@ export type PopupRequestConfig = {
   validity?: number
   uniqueIdentifierType?: NullifierType.NON_SALTED | NullifierType.SALTED
   oprfKeyId?: string
-  enableBrowserEnrollment?: boolean
 }
 
 export type PopupConfigureMessage = {
@@ -63,8 +62,6 @@ export type PopupEventMessage =
     }
   | { zkpassport: true; type: "rejected" }
   | { zkpassport: true; type: "error"; message: string }
-  | { zkpassport: true; type: "enrollment-saved" }
-  | { zkpassport: true; type: "enrollment-declined" }
 
 export type PopupMessage = PopupConfigureMessage | PopupReadyMessage | PopupEventMessage
