+++ packages/zkpassport-ui/src/card.tsx
@@ -22,7 +22,6 @@ import { injectStylesheet } from "./inject-styles"
 import { useCard, type CardState, type ProofStreamProgress } from "./use-card"
 import { describeQuery, type QueryDescriptionItem } from "./query-description"
 import type { ZKPassportQRCodeOptions } from "./types"
-import type { LocalProofProgress, SavedEnrollment } from "@zkpassport/sdk"
 
 export type CardControl = { retry: () => void }
 
@@ -38,7 +37,6 @@ export function Card({ options, controlRef }: CardProps) {
 
   const {
     state,
-    flow,
     url,
     qrSvg,
     query,
@@ -46,19 +44,7 @@ export function Card({ options, controlRef }: CardProps) {
     serviceLogo,
     retry,
     continueWithPhone,
-    storageSupported,
-    savedEnrollments,
-    pendingSaveName,
-    localProvingId,
-    localProgress,
     proofProgress,
-    rememberInBrowser,
-    setRememberInBrowser,
-    fallbackNotice,
-    verifyLocally,
-    saveEnrollment,
-    declineEnrollment,
-    removeSavedId,
   } = useCard(options)
 
   useEffect(() => {
@@ -72,36 +58,14 @@ export function Card({ options, controlRef }: CardProps) {
   const displayHeader = options.display?.header ?? true
   const displaySteps = options.display?.steps ?? true
   const displayAppLinks = options.display?.appLinks ?? true
-  const displayBrowserVerification = options.display?.browserVerification ?? true
   const headerName = options.name ?? serviceName ?? options.domain ?? ""
   const headerIcon = options.logo ?? serviceLogo ?? ""
   const appJoined =
-    state === "scanned" ||
-    state === "generating" ||
-    state === "local-proving" ||
-    state === "save-offer" ||
-    state === "success" ||
-    state === "error"
+    state === "scanned" || state === "generating" || state === "success" || state === "error"
   const hasFacematch = !!query?.facematch
-  const overlayCaption = getOverlayCaption(state, localProgress)
+  const overlayCaption = getOverlayCaption(state)
   const canRestart = state === "waiting" || state === "scanned"
-  // Only once the request URL exists (the toggle regenerates the QR)
-  const showRememberCheckbox =
-    displayBrowserVerification &&
-    !!options.enableBrowserEnrollment &&
-    storageSupported &&
-    flow === "phone" &&
-    state === "waiting"
-  // With the intro screen disabled, the local-verify entry point moves to the QR screen
-  const showInlineLocalVerify =
-    options.display?.intro === false &&
-    displayBrowserVerification &&
-    savedEnrollments.length > 0 &&
-    state === "waiting"
-  const steps =
-    flow === "browser"
-      ? buildBrowserSteps(state, localProgress)
-      : buildPhoneSteps(state, appJoined, hasFacematch, proofProgress)
+  const steps = buildPhoneSteps(state, appJoined, hasFacematch, proofProgress)
 
   return (
     <div className="zkp-card" data-state={state} data-theme={options.theme ?? "auto"}>
@@ -154,30 +118,11 @@ export function Card({ options, controlRef }: CardProps) {
           appName={headerName || "This app"}
           items={describeQuery(query)}
           loading={query === null}
-          savedEnrollments={displayBrowserVerification ? savedEnrollments : []}
           onContinue={continueWithPhone}
-          onVerifyLocally={verifyLocally}
-          onRemoveSavedId={removeSavedId}
-          localProvingId={localProvingId}
-          notice={fallbackNotice}
         />
       ) : (
         <div className="zkp-screen">
           <QrSlot state={state} qrSvg={qrSvg} caption={overlayCaption} />
-
-          {fallbackNotice && state === "waiting" ? (
-            <p className="zkp-notice">{fallbackNotice}</p>
-          ) : null}
-
-          {showInlineLocalVerify ? (
-            <button
-              type="button"
-              className="zkp-intro-continue zkp-local-verify-inline"
-              onClick={() => verifyLocally(savedEnrollments[0].id)}
-            >
-              Verify with this browser
-            </button>
-          ) : null}
 
           {state === "waiting" && url ? (
             <a className="zkp-open-app" href={url}>
@@ -185,56 +130,7 @@ export function Card({ options, controlRef }: CardProps) {
             </a>
           ) : null}
 
-          {showRememberCheckbox ? (
-            <label className="zkp-remember">
-              <input
-                type="checkbox"
-                checked={rememberInBrowser}
-                onChange={(event) =>
-                  setRememberInBrowser((event.target as HTMLInputElement).checked)
-                }
-              />
-              <span>
-                Remember in this browser next time{" "}
-                <span className="zkp-remember-hint">(encrypted, unlocked by your passkey)</span>
-              </span>
-            </label>
-          ) : null}
-
-          {state === "save-offer" ? (
-            <div className="zkp-save-offer">
-              <p className="zkp-save-offer-text">
-                Save this ID to verify without your phone next time. It stays encrypted in this
-                browser and only your passkey can unlock it.
-              </p>
-              {pendingSaveName ? (
-                <div className="zkp-id-row zkp-id-row-static">
-                  <span
-                    className="zkp-id-shield"
-                    dangerouslySetInnerHTML={{ __html: ICON_SHIELD }}
-                  />
-                  <span className="zkp-id-name">{pendingSaveName}</span>
-                  <span className="zkp-id-meta">encrypted</span>
-                </div>
-              ) : null}
-              <button
-                type="button"
-                className="zkp-intro-continue"
-                title="Your browser will ask you to confirm with Face ID, fingerprint or PIN"
-                onClick={saveEnrollment}
-              >
-                Save with passkey
-              </button>
-              <p className="zkp-save-offer-hint">
-                Your can confirm with Face ID, fingerprint or PIN.
-              </p>
-              <button type="button" className="zkp-intro-link" onClick={declineEnrollment}>
-                Not now
-              </button>
-            </div>
-          ) : null}
-
-          {displaySteps && steps.length > 0 && state !== "save-offer" ? (
+          {displaySteps && steps.length > 0 ? (
             <>
               <div className="zkp-divider zkp-divider-top" />
 
@@ -248,7 +144,7 @@ export function Card({ options, controlRef }: CardProps) {
             </>
           ) : null}
 
-          {displayAppLinks && flow === "phone" ? (
+          {displayAppLinks ? (
             <div className={`zkp-collapse${appJoined ? " zkp-collapse-out" : ""}`}>
               <div className="zkp-collapse-inner">
                 <div className="zkp-divider zkp-divider-bottom" />
@@ -284,25 +180,13 @@ function IntroSection({
   appName,
   items,
   loading,
-  savedEnrollments,
   onContinue,
-  onVerifyLocally,
-  onRemoveSavedId,
-  localProvingId,
-  notice,
 }: {
   appName: string
   items: QueryDescriptionItem[]
   loading: boolean
-  savedEnrollments: SavedEnrollment[]
   onContinue: () => void
-  onVerifyLocally: (enrollmentId: string) => void
-  onRemoveSavedId: (enrollmentId: string) => void
-  localProvingId: string | null
-  notice: string | null
 }) {
-  const hasSaved = savedEnrollments.length > 0
-
   return (
     <div className="zkp-intro">
       <div className="zkp-intro-request">
@@ -334,91 +218,23 @@ function IntroSection({
         )}
       </div>
 
-      {hasSaved ? (
-        <>
-          <div className="zkp-divider" />
-          <div className="zkp-intro-saved">
-            <div className="zkp-saved-heading">
-              <p className="zkp-eyebrow">Saved in this browser</p>
-              <p className="zkp-saved-hint">Encrypted. Only your passkey can unlock them.</p>
-            </div>
-            <div className="zkp-id-list">
-              {savedEnrollments.map((enrollment) => {
-                const proving = localProvingId === enrollment.id
-                const disabled = localProvingId !== null
-                return (
-                  <div key={enrollment.id} className="zkp-id-row" data-proving={proving}>
-                    <button
-                      type="button"
-                      className="zkp-id-main"
-                      disabled={disabled}
-                      onClick={() => onVerifyLocally(enrollment.id)}
-                    >
-                      <span
-                        className="zkp-id-shield"
-                        dangerouslySetInnerHTML={{ __html: ICON_SHIELD }}
-                      />
-                      <span className="zkp-id-name">{enrollment.maskedName ?? "Saved ID"}</span>
-                    </button>
-                    {proving ? <span className="zkp-id-wait" aria-hidden="true" /> : null}
-                    {!proving ? (
-                      <button
-                        type="button"
-                        className="zkp-id-remove"
-                        aria-label={`Remove ${enrollment.maskedName ?? "saved ID"}`}
-                        title="Remove"
-                        disabled={disabled}
-                        onClick={() => {
-                          if (
-                            window.confirm(
-                              `Remove ${enrollment.maskedName ?? "this saved ID"} from this browser? You'll need your phone to verify and save it again.`,
-                            )
-                          ) {
-                            onRemoveSavedId(enrollment.id)
-                          }
-                        }}
-                      >
-                        ×
-                      </button>
-                    ) : null}
-                  </div>
-                )
-              })}
-            </div>
-            {notice ? <p className="zkp-intro-notice">{notice}</p> : null}
-            <button
-              type="button"
-              className="zkp-intro-secondary"
-              disabled={localProvingId !== null}
-              onClick={onContinue}
-            >
-              Use another ID
-            </button>
-          </div>
-        </>
-      ) : (
-        <>
-          <div className="zkp-intro-question">
-            <span
-              className="zkp-chip-symbol"
-              aria-hidden="true"
-              dangerouslySetInnerHTML={{ __html: ICON_EPASSPORT_CHIP }}
-            />
-            <p className="zkp-intro-question-text">
-              Do you have a passport or ID card with a chip?
-            </p>
-            <p className="zkp-intro-question-hint">
-              You can verify with a chip-enabled passport or ID without sharing your document. Your
-              phone can read it over NFC.
-              <br />
-              Look for this symbol on the cover.
-            </p>
-            <button type="button" className="zkp-intro-continue" onClick={onContinue}>
-              Continue
-            </button>
-          </div>
-        </>
-      )}
+      <div className="zkp-intro-question">
+        <span
+          className="zkp-chip-symbol"
+          aria-hidden="true"
+          dangerouslySetInnerHTML={{ __html: ICON_EPASSPORT_CHIP }}
+        />
+        <p className="zkp-intro-question-text">Do you have a passport or ID card with a chip?</p>
+        <p className="zkp-intro-question-hint">
+          You can verify with a chip-enabled passport or ID without sharing your document. Your
+          phone can read it over NFC.
+          <br />
+          Look for this symbol on the cover.
+        </p>
+        <button type="button" className="zkp-intro-continue" onClick={onContinue}>
+          Continue
+        </button>
+      </div>
 
       <div className="zkp-intro-footer">
         <div className="zkp-divider" />
@@ -445,17 +261,12 @@ function IntroSection({
 type StepStatus = "pending" | "current" | "done"
 type StepDef = { key: string; label: ComponentChildren; status: StepStatus; icon: string | null }
 
-function getOverlayCaption(state: CardState, localProgress: LocalProofProgress | null): string {
+function getOverlayCaption(state: CardState): string {
   switch (state) {
     case "scanned":
       return "Approve the request on your phone"
     case "generating":
       return "Generating proof…"
-    case "local-proving":
-      return localProgress
-        ? `Verifying in this browser (${localProgress.current}/${localProgress.total})…`
-        : "Verifying in this browser…"
-    case "save-offer":
     case "success":
       return "Verification complete"
     case "error":
@@ -471,7 +282,7 @@ function buildPhoneSteps(
   hasFacematch: boolean,
   proofProgress: ProofStreamProgress,
 ): StepDef[] {
-  const finished = state === "success" || state === "save-offer"
+  const finished = state === "success"
   const generating = state === "generating"
   const { received, total } = proofProgress
   const allReceived = total !== null && received >= total
@@ -544,34 +355,6 @@ function buildPhoneSteps(
   return appJoined ? steps : steps.slice(0, 3)
 }
 
-function buildBrowserSteps(state: CardState, localProgress: LocalProofProgress | null): StepDef[] {
-  const finished = state === "success" || state === "save-offer"
-  const proving = state === "local-proving"
-  return [
-    {
-      key: "unlock",
-      label: "Unlock your saved ID with your passkey.",
-      status:
-        proving && !localProgress ? "current" : localProgress || finished ? "done" : "pending",
-      icon: null,
-    },
-    {
-      key: "generate",
-      label: localProgress
-        ? `Generate proof in this browser (${localProgress.current}/${localProgress.total}).`
-        : "Generate proof in this browser.",
-      status: proving && localProgress ? "current" : finished ? "done" : "pending",
-      icon: null,
-    },
-    {
-      key: "verify",
-      label: "Verify the proof.",
-      status: finished ? "done" : "pending",
-      icon: null,
-    },
-  ]
-}
-
 function QrSlot({
   state,
   qrSvg,
@@ -581,7 +364,7 @@ function QrSlot({
   qrSvg: string | null
   caption: string
 }) {
-  const showSpinner = state === "connecting" || state === "generating" || state === "local-proving"
+  const showSpinner = state === "connecting" || state === "generating"
 
   return (
     <div className="zkp-qr-slot" data-state={state}>
@@ -598,7 +381,7 @@ function QrSlot({
           {state === "scanned" ? (
             <div className="zkp-scanned-phone" dangerouslySetInnerHTML={{ __html: ICON_PHONE }} />
           ) : null}
-          {state === "success" || state === "save-offer" ? (
+          {state === "success" ? (
             <div className="zkp-check" dangerouslySetInnerHTML={{ __html: ICON_CHECK }} />
           ) : null}
           {state === "error" ? (
