+++ packages/zkpassport-utils/src/index.ts
@@ -2,7 +2,6 @@ export * from "./barrett-reduction"
 export * from "./binary"
 export * from "./circuit-matcher"
 export * from "./circuits"
-export * from "./enrollment"
 export * from "./cms"
 export * from "./constants"
 export * from "./merkle-tree"
