+++ packages/zkpassport-utils/src/types/index.ts
@@ -427,22 +427,6 @@ export type DisclosureWitness = {
   salt: string
 }
 
-/**
- * Bundle sent by the mobile app over the bridge (message method "enrollment") that lets
- * a browser generate disclosure proofs locally on future verification requests.
- */
-export type EnrollmentBundle = {
-  // Bundle format version
-  version: 1
-  // The circuit manifest version the base subproofs were generated against
-  circuitVersion: `${number}.${number}.${number}`
-  // 0x-prefixed hex root of the certificate registry the DSC subproof was built against
-  certificateRegistryRoot: string
-  // Exactly 3 proofs: sig_check_dsc*, sig_check_id_data*, data_check_integrity*
-  baseSubproofs: ProofResult[]
-  witness: DisclosureWitness
-}
-
 export type ProofMode = "fast" | "compressed" | "compressed-evm"
 
 export type QRCodeData = {
@@ -458,9 +442,6 @@ export type QRCodeData = {
   uniqueIdentifierType: NullifierType | null
   oprfKeyId: string | null
   returnDeepLink: string | null
-  // Whether the requesting browser wants to receive an enrollment bundle
-  // after a successful verification (URL param `be=1`)
-  enrollment?: boolean | null
 }
 
 export interface JsonRpcRequest {
